GUI: add gui.experimental and move experimental components here Add support for package level doc string in quickindex
