+++ arcade/gui/examples/__init__.py
@@ -1,0 +1,7 @@
+"""
+This package contains GUI specific examples.
+
+These show explicit GUI examples, and are not part of the main examples.
+
+Examples which begin with "exp_" use experimental components from the `arcade.gui.experimental` module.
+"""
