+++ arcade/gui/experimental/__init__.py
@@ -1,0 +1,15 @@
+"""
+Contains experimental GUI elements.
+
+The API of these components may change within minor version updates.
+No Deprecation warnings are given for changes in this module.
+"""
+from arcade.gui.examples.textured_slider import UITextureSlider
+from arcade.gui.experimental.scroll_area import UIScrollArea
+from arcade.gui.experimental.password_input import UIPasswordInput
+
+__all__ = [
+    "UITextureSlider",
+    "UIScrollArea",
+    "UIPasswordInput",
+]
