+++ arcade/gui/experimental/scroll_area.py
@@ -1,19 +1,3 @@
-"""
-This example is a proof-of-concept for a UIScrollArea.
-
-You can currently scroll through the UIScrollArea in the following ways:
-
-* scrolling the mouse wheel
-* dragging with middle mouse button
-
-It currently needs the following improvements:
-
-* A better API, including scroll direction control
-* UIScrollBars
-
-If arcade and Python are properly installed, you can run this example with:
-python -m arcade.gui.examples.scroll_area
-"""
 from __future__ import annotations
 
 from typing import Iterable, Optional
@@ -21,28 +5,52 @@ from typing import Iterable, Optional
 from pyglet.event import EVENT_UNHANDLED
 
 import arcade
-from arcade import Window
-from arcade.gui import UIManager, UIWidget, Property, Surface, UIDummy, UIEvent, bind, \
-    UIMouseDragEvent, UIMouseScrollEvent, UIMouseEvent, UIBoxLayout, UIFlatButton, UIInputText
+from arcade.gui import (
+    UIWidget,
+    Property,
+    Surface,
+    bind,
+    UIEvent,
+    UIMouseDragEvent,
+    UIMouseScrollEvent,
+    UIMouseEvent,
+)
 
 
 class UIScrollArea(UIWidget):
+    """A widget that can scroll its children."""
+
     scroll_x = Property[float](default=0.0)
     scroll_y = Property[float](default=0.0)
 
     scroll_speed = 1.3
     invert_scroll = False
 
-    def __init__(self, *,
-                 x: float = 0, y: float = 0, width: float = 300, height: float = 300,
-                 children: Iterable["UIWidget"] = tuple(),
-                 size_hint=None,
-                 size_hint_min=None,
-                 size_hint_max=None,
-                 canvas_size=(300, 300),
-                 **kwargs):
-        super().__init__(x=x, y=y, width=width, height=height, children=children, size_hint=size_hint,
-                         size_hint_min=size_hint_min, size_hint_max=size_hint_max, **kwargs)
+    def __init__(
+        self,
+        *,
+        x: float = 0,
+        y: float = 0,
+        width: float = 300,
+        height: float = 300,
+        children: Iterable["UIWidget"] = tuple(),
+        size_hint=None,
+        size_hint_min=None,
+        size_hint_max=None,
+        canvas_size=(300, 300),
+        **kwargs,
+    ):
+        super().__init__(
+            x=x,
+            y=y,
+            width=width,
+            height=height,
+            children=children,
+            size_hint=size_hint,
+            size_hint_min=size_hint_min,
+            size_hint_max=size_hint_max,
+            **kwargs,
+        )
         self.surface = Surface(
             size=canvas_size,
         )
@@ -81,12 +89,16 @@ class UIScrollArea(UIWidget):
         self.surface.draw((0, 0, width, height))
 
     def on_event(self, event: UIEvent) -> Optional[bool]:
-
-        if isinstance(event, UIMouseDragEvent) and not self.rect.collide_with_point(event.x, event.y):
+        if isinstance(event, UIMouseDragEvent) and not self.rect.collide_with_point(
+            event.x, event.y
+        ):
             return EVENT_UNHANDLED
 
         # drag scroll area around with middle mouse button
-        if isinstance(event, UIMouseDragEvent) and event.buttons & arcade.MOUSE_BUTTON_MIDDLE:
+        if (
+            isinstance(event, UIMouseDragEvent)
+            and event.buttons & arcade.MOUSE_BUTTON_MIDDLE
+        ):
             self.scroll_x -= event.dx
             self.scroll_y -= event.dy
             return True
@@ -105,30 +117,3 @@ class UIScrollArea(UIWidget):
             child_event.y = event.y - self.y + self.scroll_y
 
         return super().on_event(child_event)
-
-
-class MyWindow(Window):
-
-    def __init__(self):
-        super().__init__()
-
-        self.ui = UIManager()
-        self.ui.enable()
-        self.background_color = arcade.color.WHITE
-        self.input = self.ui.add(UIInputText(x=450, y=300).with_border())
-
-        self.scroll_area = UIScrollArea(x=100, y=100).with_border()
-        self.ui.add(self.scroll_area)
-
-        anchor = self.scroll_area.add(UIBoxLayout(width=300, height=300, space_between=20))
-        anchor.add(UIDummy(height=50))
-        anchor.add(UIFlatButton(text="Hello from scroll area", multiline=True))
-        anchor.add(UIInputText().with_border())
-
-    def on_draw(self):
-        arcade.start_render()
-        self.ui.draw()
-
-
-if __name__ == "__main__":
-    MyWindow().run()
